// src/app/gateway-model/model/gateway-model.ts
export class GatewayModel {

  id: number;
  name: string;
  description: string;
  gatewayTypeName: string;
  gatewayTypeId: String;
  isGeneric: boolean;
  status: string;
  updatedBy: number;
  updatedOn: Date;
  createdBy: number;
  createdOn: Date;
  isAvailable: string;
  businessEntityId: number;
}